[Refactor] Move database AST classes from fe-core to fe-parser
- Move CreateDbStmt, DropDbStmt, UseDbStmt from fe-core to fe-parser
- Update AstVisitor interfaces: move visit methods from AstVisitorExtendInterface to AstVisitor
- Update accept methods to call specific visit methods instead of casting to AstVisitorExtendInterface
- Improve module separation and reduce dependencies between fe-core and fe-parser
- Maintain proper AST visitor pattern implementation
