+++ fe/fe-parser/src/main/java/com/starrocks/sql/ast/CreateDbStmt.java
@@ -66,6 +66,6 @@ public class CreateDbStmt extends DdlStmt {
 
     @Override
     public <R, C> R accept(AstVisitor<R, C> visitor, C context) {
-        return ((AstVisitorExtendInterface<R, C>) visitor).visitCreateDbStatement(this, context);
+        return visitor.visitCreateDbStatement(this, context);
     }
 }
